+++ package.json
@@ -41,8 +41,10 @@
     "co": "^4.6.0"
   },
   "devDependencies": {
+    "bluebird": "^3.4.7",
     "express": "^4.14.0",
     "mocha": "^3.0.2",
+    "q": "^1.4.1",
     "should": "^10.0.0",
     "supertest": "^2.0.0"
   }
